+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [5.10.1](https://github.com/plivo/plivo-java/tree/v5.10.1) (2022-04-12)
 **Feature - List all recordings**
 - `fromNumber` and `toNumber` added to filtering param [List all recordings](https://www.plivo.com/docs/voice/api/recording#list-all-recordings)
+- `recordMinMemberCount` param added in [Add a participant to a multiparty call using API](https://www.plivo.com/docs/voice/api/multiparty-call/participants#add-a-participant)
 
 ## [5.10.0](https://github.com/plivo/plivo-java/tree/v5.10.0) (2022-03-25)
 **Feature - DialElement**